requirements.txt
packaging==23.0
Pillow==9.4.0
pytesseract==0.3.10
